- I added the 2 test cases of scrolling in one test class - I removed the test class of "VerifyingScrollUpWithoutUsingArrowButtonAndScrollDownFunctionality"
